fix : fix data type mismatch QUZ-47

+++ search-service/search-domain/src/main/kotlin/com/grepp/quizy/search/domain/quiz/SearchedQuiz.kt
@@ -107,7 +107,7 @@ data class SearchedQuizOption(val no: Int, val content: String, val selectionRat
             SearchedQuizOption(
                 option.optionNumber,
                 option.content,
-                (option.selectionCount / total).toDouble(),
+                option.selectionCount.toDouble() / total.toDouble()
             )
     }
 }
